requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
langchain_google_genai
